webscrp update
added remaining char count
fixed up some bugs

+++ WEBSCRP/612136/scripts/create_database.php
@@ -85,7 +85,7 @@
 				//create the items table
 				// add end date?
 				$query = "CREATE TABLE `items`(`itemID` int(10) NOT NULL AUTO_INCREMENT, `itemName` varchar(30), `itemQuantity` int(10),
-					`itemPrice` float(10), `sellerName` varchar(50), `isNew` int(1), 
+					`itemPrice` decimal(10, 2), `sellerName` varchar(50), `isNew` int(1), 
 					`tags` varchar(200), `itemDescription` varchar(1000), `image` varchar(200), `searchRelevance` int(100), `catagoryID` int(10),
 					PRIMARY KEY(`itemID`))";
 				// FOREIGN KEY(`catagoryID`) REFERENCES catagories(`catagoryID`)?
